Añadido el Registrarse y acomodado de detalles

+++ templates_c/dfd29fcb7df8b505db473606adeb8efd654cb264_0.file.tablaCategorias.tpl.php
@@ -1,30 +1,32 @@
 <?php
-/* Smarty version 4.2.1, created on 2022-10-15 01:18:45
+/* Smarty version 4.2.1, created on 2022-10-17 18:36:28
   from 'C:\xampp\htdocs\MotorSportPage\app\Templates\tablaCategorias.tpl' */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '4.2.1',
-  'unifunc' => 'content_6349ee553b0b47_67525545',
+  'unifunc' => 'content_634d848ca833e2_42808548',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
     'dfd29fcb7df8b505db473606adeb8efd654cb264' => 
     array (
       0 => 'C:\\xampp\\htdocs\\MotorSportPage\\app\\Templates\\tablaCategorias.tpl',
-      1 => 1665789349,
+      1 => 1666024586,
       2 => 'file',
     ),
   ),
   'includes' => 
   array (
+    'file:app/Templates/form_categoria.tpl' => 1,
+    'file:app/Templates/footer.tpl' => 1,
   ),
 ),false)) {
-function content_6349ee553b0b47_67525545 (Smarty_Internal_Template $_smarty_tpl) {
+function content_634d848ca833e2_42808548 (Smarty_Internal_Template $_smarty_tpl) {
 ?><section>
-<h1 class="text-center shadow-sm p-3 mb-0 mt-5 bg-body rounded"><?php echo $_smarty_tpl->tpl_vars['titulo_categorias']->value;?>
+<h1 class="text-center shadow-sm p-3 mb-4 mt-5 bg-body rounded"><?php echo $_smarty_tpl->tpl_vars['titulo']->value;?>
 </h1>
-<?php if (!(isset($_SESSION['ID_USUARIO']))) {?>
+<?php if (!(isset($_SESSION['ID_LOGGED']))) {?>
     <p class="p-5 fs-5">En esta tabla mostraremos los autos descriptos en la tabla anterior y la categoria competitiva a la
         que pertenecen.Mostraremos el nombre de la categoria, una detallada descripcion de su organizacion y obejtivo, y en
         que parte del mundo rige</p>
@@ -36,12 +38,10 @@ function content_6349ee553b0b47_67525545 (Smarty_Internal_Template $_smarty_tpl)
             <th>Tipo</th>
             <th>Filtrar</th>
             <th>Descripcion</th>
-            
+            <?php if ((isset($_SESSION['IS_LOGGED']))) {?>
                 <th>BORRAR</th>
-           
-            
                 <th>EDITAR</th>
-            
+            <?php }?>
         </tr>
     </thead>
     <tbody>
@@ -62,19 +62,25 @@ $_smarty_tpl->tpl_vars['categoria']->do_else = false;
                 </td>
                 <td><?php echo $_smarty_tpl->tpl_vars['categoria']->value->descripcion;?>
 </td>
-                
-                <td><a class="btn btn-badge text-bg-danger" href="deleteCategorias/<?php echo $_smarty_tpl->tpl_vars['categoria']->value->id_categorias;?>
+                <?php if ((isset($_SESSION['IS_LOGGED']))) {?>
+                    <td><a class="btn btn-badge text-bg-danger" href="deleteCategorias/<?php echo $_smarty_tpl->tpl_vars['categoria']->value->id_categorias;?>
 ">BORRAR</a></td>
-                
-                
-                <td><a class="btn btn-badge text-bg-warning" href="editCat/<?php echo $_smarty_tpl->tpl_vars['categoria']->value->id_categorias;?>
-">EDITAR</a></td>
-                
+                     <td><a class="btn btn-badge text-bg-warning" href="editCat/<?php echo $_smarty_tpl->tpl_vars['categoria']->value->id_categorias;?>
+">EDITAR</a>
+                    </td>
+                <?php }?>
             </tr>
         <?php
 }
 $_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?>
     </tbody>
 </table>
-</section><?php }
+</section>
+
+<?php if ((isset($_SESSION['IS_LOGGED']))) {?>
+    <?php $_smarty_tpl->_subTemplateRender("file:app/Templates/form_categoria.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
+}?>
+
+<?php $_smarty_tpl->_subTemplateRender("file:app/Templates/footer.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
+}
 }
